feat:login - new email redirect register; refactor:separate login and registration
