+++ src/services/auth/logoutService.ts
@@ -1,0 +1,11 @@
+import UserModel from "../../models/user";
+
+export const logoutService = {
+    logoutUser: async (refreshToken: string) => {
+        const user = await UserModel.findOne({ refreshToken });
+        if (user) {
+            user.refreshToken = undefined;
+            await user.save();
+        }
+    },
+}
